Maintain Tracer.CurrentSpan with NoopSpan
